pyproject.toml
[tool.poetry]
name = "wrongassign"
version = "0.1.0"
description = ""
authors = ["hotbari <138793425+hotbari@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
